RWRoute source code (#229)
* initial commit of tcl files to build data for rwroute regarding clock network

* initial commit of a delay computaion of routing resource to be used in rwroute, according to FPT19 paper

* timing model related changes from Pongstorn


* timing graph related changes for RWRoute


* initial commit of RWRoute new files


* current workaround to create pins of static nets


* minor changes and additional methods for RWRoute clock routing


* add routeDesign(), merge RouterBase and RWRoute, new ReportTimingAndWirelength, move Timer out, add timers for routing wire nets,
fix some code styling, better organized javadoc, as private as possible, etc.


* argument printConnectionSpan


* Delay estimator optimization - improves timing driven router runtime by 3%


* Optimizes the hashCode() and equals() of TimingEdge for runtime improvement.


* getLogicDelay() optimization among others.


* Add default partial routing configuration, add timing related class javadoc, concise warning info, correct typos, clean up, etc

* Fixes inconsistency of behavioral changes.  Should produce same routing result but with the optimizations in place.


* refatoring, more specific parameter names, clear iteration table headers, etc


* merge RoutableData into RoutableNode for memory efficiency


* delete unused RoutableData


* replace HashMultiSet with Map


* refactor, more clear variable names and javadoc of Routable, move delay setDelay to constructor, etc.


* move driversCounts update out of routing


* remove some unnecessary global members


* move two variables into the method


* use sink pin only for getting timing edges of connections, refator variables and methods


* queue element optimization, reprouding consistent results


* more specific variable names


* more specific variable names & method javadoc


* Correctly creates GND SitePinInsts on certain SRL16Es.


* Fixes an oversight when switching nets for pins.


* add a few more configuration options


* fix pin issues: CLK and CE of routethru FFs, SRLC32E, RAMB18E2


* minor changes


* simplify and add routeDesign APIs


* correct an oversight by going back to return the routed design instance


* forgot two helper class files


* remove test_data files


* add missing source code organization


* move timer class files to util, add missing code organization


* standardize variable and method names regarding the term "connection"


* correct some typos


* remove redundant code / parameters


* clarify a few configuration options


* correct a typo


* [cleaning up] refactor, add notes, remove redundant code, etc.


* rename a few methods


* replace the user map key with NetWrapper, refactor, adding menthod javadoc, etc.


* initial cleaning up of the GlobalSignalRouting class file


* add a blank line to the end


* remove static members in GlobalSignalRouting, refactor, add one forgotten file


* refactor to clean up clock routing, add javadoc and notes.
Tested.


* rename the class and a method


* add private


* rename a method and remove unused code


* refactor


* remove unused code


* use entrySet() when iterating over a map


* rename two helper classes


* change based on pull request comment made on Oct 12

* remove the code to not rely on the length from IntendCode

* fix NPE regarding TermInfo

* remove an unused argument of NetWrapper


* refine method javadoc


* a minor refactor to remove implicit limitations on the clock net


* refactor ClkSkewsAndRouteDelays


* remove unused code because of using nets as the users.
Tested.


* refactor two runtime tracker classes


* remove the use of node-to-sitepin map from InterconnectInfo, but derive from the device model. need to change START_TILE_COL from 52 to 61

* use BEL.canInvert to remove most hard coded invertible DSP BELPins.
Tested.


* convert taps to space for DelayEstimatorBase.java

* rename calcTimingGroupDelay  to calcNodeGroupDelay

* add a new line to the end


* add a new line to the end


* rename, refine code, add javadoc


* rename


* fix another A6 pin issue case


* minor changes on printing info


* merge routeToCentroid methods. Tested.


* remove debugging code


* make symmetric clock routing default


* fix an oversight when reporting WL for designs routed by Vivado starting from post-place DCPs


* a minor change to make the info general


* remove the ClkSkewData related experimental feature for now


* make printed info general for both RWRoute and TimingAndWirelengthReport


* remove unused net name info


* refactor


* minor changes for cleaning up


* Correct an oversight regarding CKEN and SRST pin delays.
Consistent results obtained.


* refactor, move formatString() to util/MessageGenerator


* add a check if the reference INT tile is valid. print useful info if it is not

* add compilation step for the Java code used in find_clock_route_template.tcl

* remove unsed variables


* correct method names


* add tests for RWRoute


* Update test/com/xilinx/rapidwright/rwroute/TestRWRoute.java


* add description of each test, remove one unnessary timing-driven routing test


* test git submodules


* test submodule


* change back to Eddie's repo link


* enable the test


* Update src/com/xilinx/rapidwright/timing/DelayModelSourceFromText.java


* remove a comment

+++ src/com/xilinx/rapidwright/timing/DelayEntry.java
@@ -46,7 +46,7 @@ class DelayEntry {
     /**
      * Valid configs of for this entry.
      */
-    public short  config;
+    public int  config;
 
     /**
      * Constructor with config parameter.
@@ -56,7 +56,7 @@ class DelayEntry {
      * @param d Delay of the arc.
      * @param c Valid configuration of the arc.
      */
-    public DelayEntry(String s, String f, String t, short d, short c) {
+    public DelayEntry(String s, String f, String t, short d, int c) {
         scope  = s;
         fr     = f;
         to     = t;
